requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
psycopg[binary]==3.1.19
aioredis==2.0.3
rq==1.15.0
python-dotenv==1.0.1

aiohttp[speedups]==3.9.5
parsel==1.9.1
requests-oauthlib==2.0.0
ebay_rest==0.3.3

pydantic==2.7.1
pgvector==0.2.4

pywebpush==1.14.0
discord-webhook==1.3.0

black==24.4.2
ruff==0.4.4
pytest==8.2.0
